requirements.txt
flake8==5.0.4
pytest==7.1.2
sentry-sdk==1.9.3
PyYAML==6.0
Faker==13.15.1
